pyproject.toml
[project]
name = "mcp-server-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=3.0.2",
    "langchain-mcp-adapters>=0.2.1",
    "python-dotenv>=1.2.1",
]
